+++ crates/enchudb-engine/src/lib.rs
@@ -15,6 +15,8 @@
 //! ```
 
 pub(crate) mod append_vec;
+pub(crate) mod append_bucket;
+pub(crate) mod lockfree_cylinder;
 pub(crate) mod region;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod growable_map;